Extract stale review repository path repair evidence
